Fix compiler warnings in tests

+++ fpga_interchange/dynamic_bitarray.cc
@@ -33,12 +33,12 @@ TEST_F(DynamicBitarrayTest, oneshot)
     for (size_t i = 0; i < 100; ++i) {
         std::vector<uint8_t> simple_storage;
         nextpnr::DynamicBitarray<> bitarray;
-        ASSERT_EQ(bitarray.bits_per_value(), CHAR_BIT);
+        ASSERT_EQ(bitarray.bits_per_value(), size_t(CHAR_BIT));
 
         simple_storage.resize(i);
         bitarray.resize(i);
         ASSERT_LE(i, bitarray.size());
-        ASSERT_LT(bitarray.size() - i, CHAR_BIT);
+        ASSERT_LT(bitarray.size() - i, size_t(CHAR_BIT));
 
         for (size_t k = 0; k < 3; ++k) {
             for (size_t j = 0; j < i; ++j) {
